+++ tests/fixtures/model_form_exclude.py
@@ -6,6 +6,7 @@ class User(models.ModelForm):
         model = User
         exclude = ('name',)
 
+
 class User2(ModelForm):
     name = models.CharField(max_length=255)
     email = models.EmailField()